+++ apps/backend/src/app/node/behaviors/node-behavior.function.ts
@@ -1,35 +1,30 @@
-import { Entity, OneToOne } from "@mikro-orm/core";
-import { applyDecorators } from "@nestjs/common";
+import { Entity, OneToOne, Property } from "@mikro-orm/core";
 import { NodeBehaviorFunctionDto as DTO } from "~/lib/common/app/node/dtos/behaviors";
 import { NodeBehaviorType } from "~/lib/common/app/node/dtos/behaviors/node-behavior.type";
 
 import { NodeBehaviorBase } from "./node-behavior.base";
-import { ManyToOneParams } from "../../_lib/entity/decorators";
 import { GraphEntity } from "../../graph/graph.entity";
 
+/** @internal */
+const fieldName = "__graph" satisfies keyof DTO;
 const type = NodeBehaviorType.FUNCTION;
-
-const GraphProperty = ({ foreign }: Pick<ManyToOneParams, "foreign">) =>
-	applyDecorators(
-		OneToOne(() => GraphEntity, {
-			fieldName: "__graph" satisfies keyof DTO,
-			hidden: foreign,
-			mapToPk: !foreign,
-			nullable: false,
-			onDelete: "cascade",
-			onUpdateIntegrity: "cascade",
-			owner: true,
-			persist: !foreign,
-			unique: true
-		}) as never
-	);
 
 @Entity({ discriminatorValue: type })
 export class NodeBehaviorFunction extends NodeBehaviorBase<typeof type> implements DTO {
 	/** @inheritDoc */
-	@GraphProperty({ foreign: false })
+	@Property({ fieldName, formula: alias => `${alias}.${fieldName}`, persist: false })
 	public readonly __graph!: number;
 
-	@GraphProperty({ foreign: true })
+	@OneToOne(() => GraphEntity, {
+		fieldName,
+		hidden: true,
+		mapToPk: false,
+		nullable: false,
+		onDelete: "cascade",
+		onUpdateIntegrity: "cascade",
+		owner: true,
+		persist: true,
+		unique: true
+	})
 	public readonly graph?: GraphEntity;
 }