+++ apps/backend/src/app/node/exceptions/index.ts
@@ -1,2 +1,3 @@
 export * from "./node.exception";
+export * from "./node.no-template-parameter.exception";
 export * from "./node.readonly-kind-type.exception";