+++ apps/backend/src/app/node/input/index.ts
@@ -1,3 +1,0 @@
-export * from "./node-input.entity";
-export * from "./node-input.repository";
-export * from "./node-input.types";