+++ apps/backend/src/app/node/output/index.ts
@@ -1,3 +1,0 @@
-export * from "./node-output.entity";
-export * from "./node-output.repository";
-export * from "./node-output.types";