+++ libs/common/src/app/node/dtos/input/index.ts
@@ -1,2 +1,3 @@
 export * from "./node-input.create.dto";
 export * from "./node-input.dto";
+export * from "./node-input.update.dto";