+++ libs/common/src/app/node/dtos/input/node-input.create.dto.ts
@@ -4,5 +4,7 @@ import { NodeInputDto } from "./node-input.dto";
 
 /**
  * DTO to use when creating a `node-input`
+ *
+ * Only possible for `code` behaviors.
  */
 export class NodeInputCreateDto extends PickType(NodeInputDto, ["name", "type"]) {}