+++ libs/common/src/app/node/dtos/output/index.ts
@@ -1,2 +1,3 @@
 export * from "./node-output.create.dto";
 export * from "./node-output.dto";
+export * from "./node-output.update.dto";