+++ libs/common/src/app/node/dtos/output/node-output.create.dto.ts
@@ -1,8 +1,6 @@
-import { PickType } from "@nestjs/mapped-types";
-
-import { NodeOutputDto } from "./node-output.dto";
-
 /**
- * DTO to use when creating a `node-output`
+ * DTO to use when creating a `node-output`.
  */
-export class NodeOutputCreateDto extends PickType(NodeOutputDto, ["name", "type"]) {}
+export class NodeOutputCreateDto {
+	// Can not create
+}